# Speed up `in` checks