--- a/Project.toml
+++ b/Project.toml
@@ -38,3 +38,6 @@ PlotlyExtensionsHelper = "0.2.0"
 StaticArrays = "1.9.7"
 Unitful = "1.21.0"
 julia = "1.11"
+
+[sources]
+CountriesBorders = { url = "https://github.com/JuliaSatcomFramework/CountriesBorders.jl", rev = "bc02bbd307bd619c2b421732d206f98a924351d6"}