--- a/Project.toml
+++ b/Project.toml
@@ -40,4 +40,4 @@ Unitful = "1.21.0"
 julia = "1.11"
 
 [sources]
-CountriesBorders = { url = "https://github.com/JuliaSatcomFramework/CountriesBorders.jl", rev = "bc02bbd307bd619c2b421732d206f98a924351d6"}
+CountriesBorders = { url = "https://github.com/JuliaSatcomFramework/CountriesBorders.jl", rev = "main"}