--- a/Project.toml
+++ b/Project.toml
@@ -38,6 +38,3 @@ PlotlyExtensionsHelper = "0.2.0"
 StaticArrays = "1.9.7"
 Unitful = "1.21.0"
 julia = "1.11"
-
-[sources]
-CountriesBorders = { url = "https://github.com/JuliaSatcomFramework/CountriesBorders.jl", rev = "main"}