--- a/src/interface_func.jl
+++ b/src/interface_func.jl
@@ -70,7 +70,7 @@ Meshes.centroid(d::PolyRegionOffset) = centroid(Cartesian, d.domain)
 ## CountriesBorders.extract_countries()
 CountriesBorders.extract_countries(r::GeoRegion) = r.domain
 
-## extract_plot_coords
-CountriesBorders.extract_plot_coords(b::Union{PolyBorder, MultiBorder}) = extract_plot_coords(borders(LatLon, b))
+## geom_iterable, used by extract_plot_coords
+CountriesBorders.geom_iterable(b::Union{PolyBorder, MultiBorder}) = geom_iterable(borders(LatLon, b))
 
-CountriesBorders.extract_plot_coords(r::AbstractRegion) = extract_plot_coords(r.domain)
+CountriesBorders.geom_iterable(r::AbstractRegion) = geom_iterable(r.domain)