--- a/test/Project.toml
+++ b/test/Project.toml
@@ -1,11 +1,7 @@
 [deps]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
-CountriesBorders = "ee40d4b5-31c3-4fe2-b6cc-5ac7adf7f414"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 Meshes = "eacbb407-ea5a-433e-ab97-5258b1ca43fa"
 PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestItemRunner = "f8b46487-2199-4994-9208-9a1283c18c0a"
-
-[sources]
-CountriesBorders = { url = "https://github.com/JuliaSatcomFramework/CountriesBorders.jl", rev = "main"}