This PR exploits the strategy and functionality added in https://github.com/JuliaSatcomFramework/CountriesBorders.jl/pull/47 to speed up the computations also for types of GeoGrids.

It also simplifies some of the internal functions which were quite convoluted.

For the moment it doesn't remove convexhull fields or fix some of the misplaced type parameters for GeoGrids types as that would be breaking, but it just avoid using it for checking inclusion.

Removal of hanging fields/type parameters shall be carried out at the next breaking release.

The time comparison with this PR using the following benchamrk code is shown below:

## Benchmark Code
```julia
using GeoGrids
using CoordRefSystems
using PrettyChairmarks

grid = [LatLon(lat, lon) for lat in -90:90, lon in -180:180]

tests = [
    "All Countries" => GeoRegion(; admin = "*"),
    "Europe" => GeoRegion(; continent = "Europe"),
    "Canada" => GeoRegion(; admin = "Canada"),
    "Russia" => GeoRegion(; admin = "Russia"),
    "France" => GeoRegion(; admin = "France"),
]

f(dmn, pts) = filter(in(dmn), pts)

for (nm, dmn) in tests
    println("Testing with GeoRegion with $nm")
    println()
    b = @bs f($dmn, $grid)
    display(b)
    println()
    println()
    gro = GeoRegionOffset(dmn, 100e3)
    println("Testing with GeoRegionOffset with $nm")
    println()
    b = @bs f($gro, $grid)
    display(b)
    println()
    println()
end
```

## Before PR (convexhull check prefiltering)
```julia
Testing with GeoRegion with All Countries

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 2.708 s (0.00% GC) to evaluate,
 with a memory estimate of 1.33 MiB, over 5 allocations.


Testing with GeoRegionOffset with All Countries

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 2.102 s (0.00% GC) to evaluate,
 with a memory estimate of 1.35 MiB, over 5 allocations.


Testing with GeoRegion with Europe

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 572.127 ms (0.00% GC) to evaluate,
 with a memory estimate of 1.05 MiB, over 5 allocations.


Testing with GeoRegionOffset with Europe

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 445.181 ms (0.00% GC) to evaluate,
 with a memory estimate of 1.07 MiB, over 5 allocations.


Testing with GeoRegion with Canada

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 204.152 ms (0.00% GC) to evaluate,
 with a memory estimate of 1.02 MiB, over 5 allocations.


Testing with GeoRegionOffset with Canada

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 201.396 ms (0.00% GC) to evaluate,
 with a memory estimate of 1.03 MiB, over 5 allocations.


Testing with GeoRegion with Russia

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 148.621 ms (0.00% GC) to evaluate,
 with a memory estimate of 1.04 MiB, over 5 allocations.


Testing with GeoRegionOffset with Russia

Chairmarks.Benchmark: 1 sample with 1 evaluation.
 Single result which took 141.911 ms (4.80% GC) to evaluate,
 with a memory estimate of 1.05 MiB, over 5 allocations.


Testing with GeoRegion with France

Chairmarks.Benchmark: 5 samples with 1 evaluation.
 Range (min … max):  17.919 ms … 25.116 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     19.435 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.177 ms ±  2.832 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁       ▁   █                                             ▁
  █▁▁▁▁▁▁▁█▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  17.9 ms         Histogram: frequency by time        25.1 ms <

 Memory estimate: 1022.28 KiB, allocs estimate: 4.


Testing with GeoRegionOffset with France

Chairmarks.Benchmark: 6 samples with 1 evaluation.
 Range (min … max):  16.483 ms … 20.553 ms  ┊ GC (min … max): 0.00% … 14.28%
 Time  (median):     17.239 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   17.590 ms ±  1.519 ms  ┊ GC (mean ± σ):  2.38% ±  5.83%

  █      ▁     ▁ ▁                                          ▁
  █▁▁▁▁▁▁█▁▁▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  16.5 ms         Histogram: frequency by time        20.6 ms <

 Memory estimate: 1023.16 KiB, allocs estimate: 5.
```

## After PR (no convex hull prefiltering but boundinbox prefiltering per polyarea)
```julia
Testing with GeoRegion with All Countries

Chairmarks.Benchmark: 2 samples with 1 evaluation.
 Range (min … max):  89.731 ms … 94.934 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     92.332 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   92.332 ms ±  3.679 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █                                                         █
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  89.7 ms         Histogram: frequency by time        94.9 ms <

 Memory estimate: 1.33 MiB, allocs estimate: 5.


Testing with GeoRegionOffset with All Countries

Chairmarks.Benchmark: 2 samples with 1 evaluation.
 Range (min … max):  53.096 ms … 57.640 ms  ┊ GC (min … max): 0.00% … 9.85%
 Time  (median):     55.368 ms              ┊ GC (median):    4.92%
 Time  (mean ± σ):   55.368 ms ±  3.213 ms  ┊ GC (mean ± σ):  4.92% ± 6.96%

  █                                                         █
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  53.1 ms         Histogram: frequency by time        57.6 ms <

 Memory estimate: 1.35 MiB, allocs estimate: 5.


Testing with GeoRegion with Europe

Chairmarks.Benchmark: 6 samples with 1 evaluation.
 Range (min … max):  18.292 ms … 21.624 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     19.974 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.859 ms ±  1.150 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █           █              █    █  █                      █
  █▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁█▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  18.3 ms         Histogram: frequency by time        21.6 ms <

 Memory estimate: 1.05 MiB, allocs estimate: 5.


Testing with GeoRegionOffset with Europe

Chairmarks.Benchmark: 7 samples with 1 evaluation.
 Range (min … max):  13.024 ms … 17.245 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.514 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.090 ms ±  1.642 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █ █             █                 █       █    █          █
  █▁█▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁█▁▁▁▁█▁▁▁▁▁▁▁▁▁▁█ ▁
  13 ms           Histogram: frequency by time        17.2 ms <

 Memory estimate: 1.07 MiB, allocs estimate: 5.


Testing with GeoRegion with Canada

Chairmarks.Benchmark: 17 samples with 1 evaluation.
 Range (min … max):  5.380 ms …   6.945 ms  ┊ GC (min … max): 0.00% … 17.20%
 Time  (median):     5.832 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.928 ms ± 414.657 μs  ┊ GC (mean ± σ):  1.80% ±  5.12%

              █    ▃
  ▇▁▁▁▁▁▁▁▇▁▁▇█▁▇▇▁█▇▇▁▇▁▇▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▇ ▁
  5.38 ms         Histogram: frequency by time        6.95 ms <

 Memory estimate: 1.02 MiB, allocs estimate: 5.


Testing with GeoRegionOffset with Canada

Chairmarks.Benchmark: 19 samples with 1 evaluation.
 Range (min … max):  4.791 ms … 10.180 ms  ┊ GC (min … max): 0.00% … 15.16%
 Time  (median):     5.190 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.527 ms ±  1.167 ms  ┊ GC (mean ± σ):  1.44% ±  4.34%

     █  ▁
  ▄▁▇█▄▇█▄▁▁▄▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▁
  4.79 ms        Histogram: frequency by time        10.2 ms <

 Memory estimate: 1.03 MiB, allocs estimate: 5.


Testing with GeoRegion with Russia

Chairmarks.Benchmark: 9 samples with 1 evaluation.
 Range (min … max):   9.677 ms … 15.110 ms  ┊ GC (min … max): 0.00% … 10.05%
 Time  (median):     11.145 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   11.752 ms ±  1.651 ms  ┊ GC (mean ± σ):  1.12% ±  3.35%

  █       █  ██  █            █   █ █                       █
  █▁▁▁▁▁▁▁█▁▁██▁▁█▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  9.68 ms         Histogram: frequency by time        15.1 ms <

 Memory estimate: 1.04 MiB, allocs estimate: 5.


Testing with GeoRegionOffset with Russia

Chairmarks.Benchmark: 11 samples with 1 evaluation.
 Range (min … max):  8.759 ms …  11.314 ms  ┊ GC (min … max): 0.00% … 8.84%
 Time  (median):     9.215 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.465 ms ± 718.116 μs  ┊ GC (mean ± σ):  0.80% ± 2.67%

  ▁  ▁▁ ▁   █    ▁   ▁    ▁ ▁                               ▁
  █▁▁██▁█▁▁▁█▁▁▁▁█▁▁▁█▁▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  8.76 ms         Histogram: frequency by time        11.3 ms <

 Memory estimate: 1.05 MiB, allocs estimate: 5.


Testing with GeoRegion with France

Chairmarks.Benchmark: 134 samples with 1 evaluation.
 Range (min … max):  534.700 μs …   1.727 ms  ┊ GC (min … max): 0.00% … 64.92%
 Time  (median):     605.950 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   711.776 μs ± 264.562 μs  ┊ GC (mean ± σ):  6.20% ± 17.61%

   ▃█▄▄▂▂ ▁▁
  ▇██████▆██▇▅▅▇▁▁▁▁▁▁▁▁▅▅▅▁▁▁▁▁▁▁▁▅▁▁▁▁▅▁▁▁▆▅▅▅▁▁▁▁▅▅▅▁▆▅▁▁▁▁▁ ▅
  535 μs        Histogram: log(frequency) by time       1.63 ms <

 Memory estimate: 1022.28 KiB, allocs estimate: 4.


Testing with GeoRegionOffset with France

Chairmarks.Benchmark: 178 samples with 1 evaluation.
 Range (min … max):  369.500 μs …   1.953 ms  ┊ GC (min … max): 0.00% … 72.05%
 Time  (median):     408.700 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   527.852 μs ± 302.281 μs  ┊ GC (mean ± σ):  7.13% ± 20.15%

  ▆█▅▁▁▂▁
  ███████▆▄▁▁▁▁▄▄▄▁▄▁▁▁▁▄▁▁▁▄▁▁▁▁▁▁▄▁▆▆▁▆▇▁▄▁▁▁▁▁▆▄▁▁▁▁▁▁▁▁▆▁▄▄ ▄
  370 μs        Histogram: log(frequency) by time        1.6 ms <

 Memory estimate: 1023.16 KiB, allocs estimate: 5.
```